refactor(platform): checkpoint process ownership migration (#1377)
* refactor(platform): move process ownership primitives

* refactor(platform): centralize stdio and crash mechanics

* refactor(platform): route command and entry setup

* refactor(daemon): split process policy tests

* fix(platform): unblock cross-platform validation
