+++ crates/zccache-platform/Cargo.toml
@@ -12,6 +12,7 @@ publish = false
 [dependencies]
 # Native mechanics arrive per capability phase (fs, ipc, process, executable,
 # host); the leaf must never depend on any zccache-* crate.
+crash-handler = "0.6"
 tokio = { workspace = true }
 
 [target.'cfg(windows)'.dependencies]
@@ -22,6 +23,8 @@ windows-sys = { version = "0.61", features = [
     "Win32_Storage_FileSystem",
     "Win32_System_IO",
     "Win32_System_Ioctl",
+    "Win32_System_JobObjects",
+    "Win32_System_Console",
     "Win32_System_Threading",
 ] }
