+++ crates/zccache-platform/src/platform_linux/process/command.rs
@@ -1,1 +1,3 @@
 //! Linux command setup.
+
+pub fn hide_window(_command: &mut std::process::Command) {}
