+++ crates/zccache-platform/src/platform_macos/process/command.rs
@@ -1,1 +1,3 @@
 //! macOS command setup.
+
+pub fn hide_window(_command: &mut std::process::Command) {}
