+++ crates/zccache-platform/src/platform_win/process/command.rs
@@ -1,1 +1,8 @@
 //! Windows command setup.
+
+use std::os::windows::process::CommandExt;
+
+pub fn hide_window(command: &mut std::process::Command) {
+    const CREATE_NO_WINDOW: u32 = 0x0800_0000;
+    command.creation_flags(CREATE_NO_WINDOW);
+}
